Fix nuxt ssr bundling

+++ runtimes/node/versions/latest/helpers/nuxt/bundle.sh
@@ -4,7 +4,10 @@ if [ -n "$OPEN_RUNTIMES_OUTPUT_DIRECTORY" ]; then
     cd $OPEN_RUNTIMES_OUTPUT_DIRECTORY
 fi
 
-mv /usr/local/build/package*.json ./
-mv /usr/local/build/node_modules/ ./node_modules/
 
-rm -rf ./server/node_modules
+ENTRYPOINT="./server/index.mjs"
+if [ -e "$ENTRYPOINT" ]; then
+    mv /usr/local/build/package*.json ./
+    mv /usr/local/build/node_modules/ ./node_modules/
+    rm -rf ./server/node_modules
+fi
